+++ templates/passive/network/Interesting_Domain_Found.sh
@@ -1,7 +1,8 @@
 AUTHOR='@xer0dayz'
 VULN_NAME='Interesting Domain Found'
-FILENAME="$LOOT_DIR/domains/domains-all-sorted_NA.txt"
-MATCH='admin|dev|portal|stage|prod|tst|test'
+echo "$TARGET" > /tmp/target 
+FILENAME="/tmp/target"
+MATCH="admin|dev|portal|stage|prod|tst|test"
 SEVERITY='P5 - INFO'
 GREP_OPTIONS='-i'
 SEARCH='positive'