+++ templates/passive/web/Autocomplete_Enabled.sh
@@ -1,6 +1,6 @@
 AUTHOR='@xer0dayz'
 VULN_NAME='Autocomplete Enabled'
-FILENAME="$LOOT_DIR/web/websource-$TARGET-*.txt"
+FILENAME="$LOOT_DIR/web/websource-htt*-$TARGET-*.txt"
 MATCH='autocomplete=\"on\"'
 SEVERITY='P4 - LOW'
 GREP_OPTIONS='-i'