+++ templates/passive/web/CORS_Policy_-_Allow-Credentials_Enabled.sh
@@ -1,6 +1,6 @@
 AUTHOR='@xer0dayz'
 VULN_NAME='CORS Policy - Allow-Credentials Enabled'
-FILENAME="$LOOT_DIR/web/headers-htt*-$TARGET.txt"
+FILENAME="$LOOT_DIR/web/headers-htt*-$TARGET-*.txt"
 MATCH='Access-Control-Allow-Credentials: true'
 SEVERITY='P4 - LOW'
 GREP_OPTIONS='-i'