+++ templates/passive/web/Expired_SSL_Certificate.sh
@@ -1,6 +1,6 @@
 AUTHOR='@xer0dayz'
 VULN_NAME='Expired SSL Certificate'
-FILENAME="$LOOT_DIR/web/curldebug-$TARGET.txt"
+FILENAME="$LOOT_DIR/web/curldebug-$TARGET-*.txt"
 MATCH='certificate has expired'
 SEVERITY='P3 - MEDIUM'
 GREP_OPTIONS=''