+++ templates/passive/web/Fortinet_FortiGate_SSL_VPN_Panel_Passive_Detection.sh
@@ -1,6 +1,6 @@
 AUTHOR='@xer0dayz'
 VULN_NAME='Fortinet FortiGate SSL VPN Panel Passive Detection'
-FILENAME="$LOOT_DIR/web/headers-htt*-$TARGET.txt"
+FILENAME="$LOOT_DIR/web/headers-htt*-$TARGET-*.txt"
 MATCH="Server\:\ xxxxxxxx-xxxxx"
 SEVERITY='P5 - INFO'
 GREP_OPTIONS='-i'