+++ templates/passive/web/Insecure_Cookie_-_Secure_Not_Set.sh
@@ -1,6 +1,6 @@
 AUTHOR='@xer0dayz'
 VULN_NAME='Insecure Cookie - Secure Not Set'
-FILENAME="$LOOT_DIR/web/headers-htt*-$TARGET.txt"
+FILENAME="$LOOT_DIR/web/headers-htt*-$TARGET-*.txt"
 MATCH='Set-Cookie'
 SEVERITY='P3 - MEDIUM'
 GREP_OPTIONS='-i'