+++ templates/passive/web/Insecure_SSL_TLS_Connection.sh
@@ -1,10 +1,8 @@
-if [ "$SSL" = "true" ]; then
-	AUTHOR='@xer0dayz'
-	VULN_NAME='Weak SSL TLS Protocols'
-	FILENAME="$LOOT_DIR/web/sslscan-$TARGET.txt"
-	MATCH=' SSLv'
-	SEVERITY='P2 - HIGH'
-	GREP_OPTIONS='-i'
-	SEARCH='positive'
-	SECONDARY_COMMANDS=''
-fi
+AUTHOR='@xer0dayz'
+VULN_NAME='Weak SSL TLS Protocols'
+FILENAME="$LOOT_DIR/web/sslscan-$TARGET.txt $LOOT_DIR/web/sslscan-$TARGET-*.txt"
+MATCH="SSLv*   enabled"
+SEVERITY='P2 - HIGH'
+GREP_OPTIONS='-i'
+SEARCH='positive'
+SECONDARY_COMMANDS=''