+++ templates/passive/web/Interesting_Title_Found.sh
@@ -1,7 +1,7 @@
 AUTHOR='@xer0dayz'
 VULN_NAME='Interesting Title Found'
-FILENAME="$LOOT_DIR/web/title-htt*-$TARGET.txt"
-MATCH='admin|dev|portal'
+FILENAME="$LOOT_DIR/web/title-htt*-$TARGET-*.txt"
+MATCH='admin|dev|portal|login|sign|signup|registration|account'
 SEVERITY='P5 - INFO'
 GREP_OPTIONS='-i'
 SEARCH='positive'