+++ templates/passive/web/X-Powered-By_Header_Found.sh
@@ -1,6 +1,6 @@
 AUTHOR='@xer0dayz'
 VULN_NAME='X-Powered-By Header Found'
-FILENAME="$LOOT_DIR/web/headers-htt*-$TARGET.txt"
+FILENAME="$LOOT_DIR/web/headers-htt*-$TARGET-*.txt"
 MATCH='X-Powered-By'
 SEVERITY='P5 - INFO'
 GREP_OPTIONS='-i'