+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## CHANGELOG:
+* v9.1 - Updated Nuclei sc0pe templates
+* v9.1 - Fixed issue with Nuclei sc0pe parsers not working
 * v9.1 - Fixed issue with GAU installer/commmand not working
 * v9.1 - Fixed issue with passive URL fetching 
 * v9.1 - Fixed issue with nuclei not being installed